Progressbar: Fix progress jumping
The progressbar will now dynamically adjust its speed instead of copying the current progress. This way it smoothly adjusts to smaller differences between where the song is actually according to the metadata and where the progress bar is at. If the difference is too big the bar will jump because this likely means that the user just moved the position.